Changed texture handling in C++. Fragmented the shaders for readability.

+++ include/Globals.h
@@ -8,7 +8,7 @@
 #pragma once
 extern bool flashlightOn;
 extern int  renderMode;
-extern float widthG, heightG;
+extern int widthG, heightG;
 extern const int FRAMES_IN_FLIGHT;
 
 #endif //GLOBALS_H
