+++ src/Globals.cpp
@@ -6,6 +6,6 @@
 
 bool flashlightOn = false;
 int renderMode = 1;
-float widthG = 1080;
-float heightG = 720;
+int widthG = 1080;
+int heightG = 720;
 const int FRAMES_IN_FLIGHT = 5;
